add a few more jobs for scraping mentioned slugs
More jobs means more jank when the number is low, but that's better than
having not enough jobs for when the number is high. The best approach would
be to launch an appropriate number of jobs based on the number of slugs to
scrape; I just haven't bothered to do so.