no, bun uses less memory in this case
I'd like the same file to work across both, but how you might switch
between bun:sqlite and better-sqlite3 is a mess. I'd like to write it
properly such that TypeScript is happy, but my Emacs is crashing and
I'm not in the bloody mood for it. So this will have to do.

This is still better than just checking out the commit before switching
to node.