sort slug chars in more conventional alphabetical order
This allows the rudimentary progress (next commit) to be a bit more
useful.

For currently running jobs, this means their order would be thrown off a
bit (so for each block we're filling in a different edge), but no item
will be skipped in the end. In other words, the rudimentary progress
will only start being more accurate for new blocks that haven't been
stored yet, while everything will still be stored.