Remove the external slugs mechanism
That requires loading a large number of strings and keeping them in
memory. We actually have a need to optimize for memory use here, so
using the SQLite database directly is better.

On B, before this a scraper process would take around 300MB of memory,
while afterwards it only takes about 150MB.